Implement removeOpsIfDead in PatternRewriter
'replaceOp' and 'replaceOpWithNewOp' accepted an optional list of
Values to be removed if dead. The removal functionality wasn't
implemented, though.

This patch implements the removal functionality but on Operations
instead of Values, since a Value cannot be removed from the IR unless all
the Value results of its Operation are dead.